+++ sdk/keyvault/azure-security-keyvault-certificates/src/certificate_client.cpp
@@ -66,9 +66,9 @@ CertificateClient::CertificateClient(
     std::string const& vaultUrl,
     std::shared_ptr<Core::Credentials::TokenCredential const> credential,
     CertificateClientOptions options)
-    : m_vaultUrl(vaultUrl), m_apiVersion(options.Version)
+    : m_vaultUrl(vaultUrl), m_apiVersion(options.ApiVersion)
 {
-  auto apiVersion = options.Version;
+  auto apiVersion = options.ApiVersion;
 
   std::vector<std::unique_ptr<HttpPolicy>> perRetrypolicies;
   {