+++ sdk/keyvault/azure-security-keyvault-keys/test/ut/key_client_test.cpp
@@ -32,18 +32,10 @@ TEST(KeyVaultKeyClientUnitTest, ServiceVersion)
 {
   auto credential
       = std::make_shared<Azure::Identity::ClientSecretCredential>("tenantID", "AppId", "SecretId");
-  {
-    // 7.3
-    EXPECT_NO_THROW(auto options = KeyClientOptions();
-                    KeyClient keyClient("http://account.vault.azure.net", credential, options);
-                    EXPECT_EQ(options.Version, "7.3"););
-  }
-  {
-    // arbitrary version
-    EXPECT_NO_THROW(auto options = KeyClientOptions(); options.Version = "1.0";
-                    KeyClient keyClient("http://account.vault.azure.net", credential, options);
-                    EXPECT_EQ(options.Version, "1.0"););
-  }
+  // 7.3
+  EXPECT_NO_THROW(auto options = KeyClientOptions();
+                  KeyClient keyClient("http://account.vault.azure.net", credential, options);
+                  EXPECT_EQ(options.ApiVersion, "7.3"););
 }
 
 TEST(KeyVaultKeyClientUnitTest, GetUrl)