+++ sdk/keyvault/azure-security-keyvault-certificates/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 4.1.0-beta.1 (Unreleased)
+## 4.1.0-beta.1 (2022-07-07)
 
 ### Features Added
 
@@ -8,7 +8,7 @@
 
 ### Breaking Changes
 
-- Removed ServiceVersion type, replaced with Version field in the CertificateClientOptions type.
+- Removed ServiceVersion type, replaced with ApiVersion field in the CertificateClientOptions type.
 
 ## 4.0.0 (2022-06-07)