+++ sdk/keyvault/azure-security-keyvault-keys/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 4.3.0-beta.1 (Unreleased)
+## 4.3.0-beta.1 (2022-07-07)
 
 ### Features Added
 
@@ -8,7 +8,7 @@
 
 ### Breaking Changes
 
-- Removed ServiceVersion type, replaced with Version field in the KeyClientOptions type.
+- Removed ServiceVersion type, replaced with ApiVersion field in the KeyClientOptions type.
 
 
 ## 4.2.0 (2021-10-05)